% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_predixcan.R
\name{get_cov_2}
\alias{get_cov_2}
\title{alternative method to get covariance from expression, for a certain gene, wit ha certain orde rof models}
\usage{
get_cov_2(expression, gene, models)
}
\description{
alternative method to get covariance from expression, for a certain gene, wit ha certain orde rof models
}
