% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_predixcan.R
\name{get_data_for_gene}
\alias{get_data_for_gene}
\title{assorted data to do predixcan fo ra given gene}
\usage{
get_data_for_gene(expression, fam, gene)
}
\description{
assorted data to do predixcan fo ra given gene
}
