% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_listed}
\alias{is_listed}
\title{Checks if text matches anything in a list of patterns
(An ugly convention)}
\usage{
is_listed(text, list = NULL)
}
\arguments{
\item{v}{Strings that will be transformed.}
}
\value{
Transformed strings
}
\description{
Checks if text matches anything in a list of patterns
(An ugly convention)
}
\keyword{gencode}
