% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_predixcan.R
\name{load_expression}
\alias{load_expression}
\title{Load predicted expression files fro ma folder}
\usage{
load_expression(folder, white_list = NULL, metaxcan_style = FALSE)
}
\description{
Load predicted expression files fro ma folder
}
