% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaxcan_files.R
\name{load_metaxcan_result_file}
\alias{load_metaxcan_result_file}
\title{Loads metaxcan result file}
\usage{
load_metaxcan_result_file(path)
}
\arguments{
\item{path}{filesystem path to the file}

\item{finite}{keep only finite results}
}
\value{
a dataframe with metaxcan results
}
\description{
Loads metaxcan result file
}
\keyword{metaxcan}
