% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaxcan_files.R
\name{load_metaxcan_result_folder_3}
\alias{load_metaxcan_result_folder_3}
\title{Loads all metaxcan results in a folder}
\usage{
load_metaxcan_result_folder_3(folder, remove_postfix = NULL,
  white_list = NULL, black_list = NULL, s = c("gene", "zscore", "pheno",
  "model"))
}
\arguments{
\item{folder}{path where the result files are}

\item{remove_postfix}{pattern to be removed from file names before their parsing}

\item{white_list}{List of patterns for selecting files to be loaded}

\item{black_list}{List of patterns for discarding files not to be loaded}

\item{s}{List of columns from the results file to keep}
}
\value{
Dataframe with metaxcan results
}
\description{
Loads all metaxcan results in a folder
}
\keyword{metaxcan}
