% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaxcan_files.R
\name{pheno_model_from_file_name}
\alias{pheno_model_from_file_name}
\title{Heuristic to get phenotype name and model name from a metaxcan results file name}
\usage{
pheno_model_from_file_name(text, remove_postfix = NULL)
}
\arguments{
\item{text}{String to be parsed}

\item{remove_postfix}{String to be removed from the input text before parsing}
}
\value{
list with phenotype and model
}
\description{
Heuristic to get phenotype name and model name from a metaxcan results file name
}
\keyword{metaxcan}
