% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{remove_id_from_ensemble}
\alias{remove_id_from_ensemble}
\title{Removes everything after a "."character, meant to remove id number of gencode names.
(An ugly convention)}
\usage{
remove_id_from_ensemble(v)
}
\arguments{
\item{v}{Strings that will be transformed.}
}
\value{
Transformed strings
}
\description{
Removes everything after a "."character, meant to remove id number of gencode names.
(An ugly convention)
}
\keyword{gencode}
